refactor/1697 remove args from parser + fix tests
