+++ test/commands/parsers/CommandParser.spec.js
@@ -1,6 +1,5 @@
 import ava from 'ava';
 import sinon from 'sinon';
-import _isEqual from 'lodash/isEqual';
 import _map from 'lodash/map';
 import _tail from 'lodash/tail';
 
@@ -13,9 +12,6 @@ const CALLSIGN_MOCK = 'AAL777';
 const CAF_MOCK = 'caf';
 const CVS_MOCK = 'cvs';
 const TAKEOFF_MOCK = 'to';
-const FH_COMMAND_MOCK = 'fh 180';
-const D_COMMAND_MOCK = 'd 030';
-const STAR_MOCK = 'star quiet7';
 
 const buildCommandString = (...args) => `${CALLSIGN_MOCK} ${args.join(' ')}`;
 
@@ -48,18 +44,6 @@ ava('does not throw when called without parameters', t => {
     t.notThrows(() => new CommandParser());
 });
 
-ava('#args returns one item when a system command is present', t => {
-    const model = new CommandParser(TIMEWARP_50_MOCK);
-
-    t.true(_isEqual(model.args, [50]));
-});
-
-ava('#args an array for each command with arg values when a transmit command is present', t => {
-    const commandStringMock = buildCommandString(FH_COMMAND_MOCK, D_COMMAND_MOCK, STAR_MOCK);
-    const model = new CommandParser(commandStringMock);
-
-    t.true(model.args.length === 3);
-});
 
 ava('sets #command with the correct name when provided a system command', t => {
     const model = new CommandParser(TIMEWARP_50_MOCK);
@@ -128,45 +112,4 @@ ava('._isSystemCommand() returns true if callsignOrTopLevelCommandName exists wi
     const model = new CommandParser(systemCommandMock);
 
     t.true(model._isSystemCommand(systemCommandMock));
-});
-
-// specific use case tests
-ava('when passed hold LAM it creates the correct command with the correct arguments', t => {
-    const commandStringMock = buildCommandString('hold', 'LAM');
-    const model = new CommandParser(commandStringMock);
-
-    t.true(model.args[0][0] === 'hold');
-    t.true(model.args[0][1] === null);
-    t.true(model.args[0][2] === null);
-    t.true(model.args[0][3] === 'lam');
-});
-
-ava('when passed dct WHAMY it creates the correct command with the correct arguments', t => {
-    const commandStringMock = buildCommandString('dct', 'WHAMY');
-    const model = new CommandParser(commandStringMock);
-
-    t.true(model.args[0][0] === 'direct');
-    t.true(model.args[0][1] === 'whamy');
-});
-
-ava('when passed dct TOU it creates the correct command with the correct arguments', t => {
-    const commandStringMock = buildCommandString('dct', 'TOU');
-    const model = new CommandParser(commandStringMock);
-
-    t.true(model.args[0][0] === 'direct');
-    t.true(model.args[0][1] === 'tou');
-});
-
-ava('when passed dct TOR it creates the correct command with the correct arguments', t => {
-    const commandStringMock = buildCommandString('dct', 'TOR');
-    const model = new CommandParser(commandStringMock);
-
-    t.true(model.args[0][0] === 'direct');
-    t.true(model.args[0][1] === 'tor');
-});
-
-ava('provides a default value for the timewarp command when no args are passed', (t) => {
-    const model = new CommandParser('timewarp');
-
-    t.true(model.args[0] === 1);
 });
